@@ -54,7 +54,8 @@ class NetworkServiceWithErrorTest: XCTestCase {
         //Then
         XCTAssertEqual(train.name, self.trainName)
         XCTAssertEqual(response, .defaultMock)
-        XCTAssertEqual(networkAccess.request?.url?.absoluteString, "https://bahn.de/train")
+        let request = await networkAccess.request
+        XCTAssertEqual(request?.url?.absoluteString, "https://bahn.de/train")
     }
 
     func testRequest_withError() async {
